+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - Fix snackbar placement.
+- Fix small range, float sliders.
 - Clean up dependencies from migration.
 - Fix small dif between sliders in shaders.